Extract backend provider test support
